+++ packages/cli/src/dep-check.ts
@@ -21,6 +21,7 @@ export function rnxDepCheck(
       ],
       ["customProfiles", "excludePackages", "init", "setVersion", "vigilant"]
     ),
+    loose: Boolean(args.loose),
     write: Boolean(args.write),
     "package-json": argv[0],
   });
@@ -44,6 +45,11 @@ export const rnxDepCheckCommand = {
     {
       name: "--init [app|library]",
       description: "Writes an initial kit config",
+    },
+    {
+      name: "--loose",
+      description:
+        "Determines how strict the React Native version requirement should be. Useful for apps that depend on a newer React Native version than their dependencies declare support for.",
     },
     {
       name: "--set-version [versions]",