+++ packages/dep-check/test/check.test.ts
@@ -13,6 +13,8 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
   const consoleLogSpy = jest.spyOn(global.console, "log");
   const consoleWarnSpy = jest.spyOn(global.console, "warn");
 
+  const defaultOptions = { loose: false, write: false };
+
   const mockManifest = {
     name: "@rnx-kit/dep-check",
     version: "0.0.1",
@@ -37,7 +39,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
   });
 
   test("returns error code when reading invalid manifests", () => {
-    expect(checkPackageManifest("package.json")).not.toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).not.toBe(0);
     expect(consoleErrorSpy).toBeCalledTimes(1);
   });
 
@@ -47,7 +49,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       dependencies: { "react-native-linear-gradient": "0.0.0" },
     });
 
-    const options = { uncheckedReturnCode: -1 };
+    const options = { ...defaultOptions, uncheckedReturnCode: -1 };
     expect(checkPackageManifest("package.json", options)).toBe(-1);
     expect(consoleWarnSpy).toBeCalled();
   });
@@ -65,7 +67,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
     });
     rnxKitConfig.__setMockConfig({ reactNativeVersion: "0.64.0" });
 
-    expect(checkPackageManifest("package.json")).toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).toBe(0);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleWarnSpy.mock.calls).toMatchSnapshot();
   });
@@ -77,7 +79,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
     });
     rnxKitConfig.__setMockConfig({ reactNativeVersion: "^0.63.0 || ^0.64.0" });
 
-    expect(checkPackageManifest("package.json")).toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).toBe(0);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleWarnSpy.mock.calls).toMatchSnapshot();
   });
@@ -86,7 +88,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
     fs.__setMockContent(mockManifest);
     rnxKitConfig.__setMockConfig({ reactNativeVersion: "0.64.0" });
 
-    expect(checkPackageManifest("package.json")).toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).toBe(0);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleWarnSpy).not.toBeCalled();
   });
@@ -106,7 +108,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       capabilities: ["core-ios"],
     });
 
-    expect(checkPackageManifest("package.json")).toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).toBe(0);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleLogSpy).not.toBeCalled();
     expect(consoleWarnSpy).not.toBeCalled();
@@ -132,7 +134,9 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       capabilities: ["core-ios"],
     });
 
-    expect(checkPackageManifest("package.json", { write: true })).toBe(0);
+    expect(
+      checkPackageManifest("package.json", { loose: false, write: true })
+    ).toBe(0);
     expect(didWriteToPath).toBe(false);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleLogSpy).not.toBeCalled();
@@ -146,7 +150,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       capabilities: ["core-ios"],
     });
 
-    expect(checkPackageManifest("package.json")).not.toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).not.toBe(0);
     expect(consoleErrorSpy).toBeCalledTimes(1);
     expect(consoleWarnSpy).not.toBeCalled();
     expect(consoleLogSpy).toBeCalledTimes(2);
@@ -164,7 +168,9 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       capabilities: ["core-ios"],
     });
 
-    expect(checkPackageManifest("package.json", { write: true })).toBe(0);
+    expect(
+      checkPackageManifest("package.json", { loose: false, write: true })
+    ).toBe(0);
     expect(didWriteToPath).toBe("package.json");
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleWarnSpy).not.toBeCalled();
@@ -186,7 +192,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       capabilities: ["core-ios"],
     });
 
-    expect(checkPackageManifest("package.json")).toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).toBe(0);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleLogSpy).not.toBeCalled();
     expect(consoleWarnSpy).not.toBeCalled();
@@ -208,7 +214,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       capabilities: ["core-ios"],
     });
 
-    expect(checkPackageManifest("package.json")).toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).toBe(0);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleLogSpy).not.toBeCalled();
     expect(consoleWarnSpy).not.toBeCalled();
@@ -230,7 +236,7 @@ describe("checkPackageManifest({ kitType: 'library' })", () => {
       capabilities: ["core-ios"],
     });
 
-    expect(checkPackageManifest("package.json")).toBe(0);
+    expect(checkPackageManifest("package.json", defaultOptions)).toBe(0);
     expect(consoleErrorSpy).not.toBeCalled();
     expect(consoleLogSpy).not.toBeCalled();
     expect(consoleWarnSpy).not.toBeCalled();